+++ oatpp-server/App.cpp
@@ -31,9 +31,11 @@ void run(boost::asio::io_context& io_context) {
   CommonModule commonModule(io_context, canRequestManager);
 
   ControlModule controlModule(io_context, canRequestManager); 
+  CoreModule coreModule(io_context, canRequestManager); 
+  SensorModule sensorModule(io_context, canRequestManager); 
 
   auto myController = std::make_shared<MyController>(
-      contentMappers, io_context, systemModule, commonModule, controlModule, canRequestManager
+      contentMappers, io_context, systemModule, commonModule, controlModule, coreModule, sensorModule, canRequestManager
   );  
   router->addController(myController);