+++ Changelog.md
@@ -1,6 +1,23 @@
 # Changelog SMBR Api Server
 The version number consists of MAJOR.MINOR identifiers. The MINOR version changes correspond to the addition of new API endpoints.
 
+# 0.15 Sensor
+- Added endpoint GET bottle/temperature
+- Added endpoint GET top and bottom bottle/measured_temperature
+- Added endpoint GET top and bottom bottle/sensor_temperature
+
+# 0.14 Core
+- Added endpoint GET /core/sid
+- Added endpoint GET /core/ip_address
+- Added endpoint GET /core/hostname
+- Added endpoint GET /core/serial
+- Added endpoint GET /core/supply/type
+- Added endpoint GET /core/supply/5v
+- Added endpoint GET /core/supply/vin
+- Added endpoint GET /core/supply/poe
+- Added endpoint GET /core/supply/current
+- Added endpoint GET /core/supply/power_draw
+
 # 0.13 Mixer
 - Added endpoint GET and POST /control/mixer/speed
 - Added endpoint GET and POST /control/mixer/flowrate